requirements.txt
openai
python-decouple
python-facebook-api
opencv-contrib-python
